requirements.txt
discord.py >= 2.5.2
dotenv >= 0.9.9
pytz >= 2025.2